Make firmware download manual
Instead of integrating it with the build process, which is tricky:

python-validity contains functionality to recognize the available fp
sensor hardware, download the correct firmware, and upload it to the
sensor.

Hardware recognition and download is implemented by the
bin/validity-sensors-firmware script and it is actually meant to be
executed during package installation on distributions like debian.

However, the hardware recognition does not seem to work in a Nix build
environment (likely due to available permissions/capabilities).
Furthermore, it is probably not good to introduce these sort of
side-effects when building a package.

There are two ways to resolve this:

1. The package needs to accept the fp sensor id as a parameter and
   downloads the firmware during the build process.
   The hardware detection of python-validity is not executed.

2. The firmware download is not part of the package build process.
   Instead, the user has to run bin/validity-sensors-firmware
   manually, the first time they use the software.

   Alternatively, we patch the python-validity service, so that it runs
   the script, if no firmware is already present.

I opted for the second option, since the first option is prone to user
error. Also, it makes the flake harder to maintain because if the
firmware recognition and download code changes in python-validity, we
have to update the package input options, too.